+++ R/zzz.R
@@ -3,8 +3,8 @@
 #' RconTROLL package global options
 #'  
 #' @param name char. name of simulations
-#' @param src char. path to the TROLL C++ main code (default: package extdata folder)
-#' @param app char. path to the TROLL executable (default: package extdata folder)
+#' @param src char. path to the TROLL C++ main code (default: package tools folder)
+#' @param app char. path to the TROLL executable (default: package tools folder)
 #' @param init char. path to the input file template (default: package extdata folder)
 #' @param species char. path to the species input data (default: package extdata folder)
 #' @param species char. path to the species input data (default: package extdata folder)
@@ -17,10 +17,8 @@ NULL
   op <- options()
   op.RconTROLL <- list(
     RconTROLL.name = "TROLL_test",
-    RconTROLL.src = system.file("tools", "main_v2.3.2.cpp",  
-                            package = 'RconTROLL'),
-    RconTROLL.app = system.file("extdata", "RconTROLL.exe",  
-                                 package = 'RconTROLL'),
+    RconTROLL.src = file.path(system.file("tools", package = 'RconTROLL'), "main_v2.3.2.cpp"),
+    RconTROLL.app = file.path(system.file("tools", package = 'RconTROLL'), "TROLL.exe"),
     RconTROLL.init = system.file("extdata", "init.txt",  
                                  package = 'RconTROLL'),
     RconTROLL.species = system.file("extdata", "species.txt",